Revert "Moved BatchInfo from streaming.scheduler to streaming.ui"
This reverts commit 5c80919dc1c03bccf997f71136d734de4e0cddfa.
