+++ streaming/src/main/scala/org/apache/spark/streaming/scheduler/StreamingListener.scala
@@ -20,7 +20,6 @@ package org.apache.spark.streaming.scheduler
 import scala.collection.mutable.Queue
 
 import org.apache.spark.util.Distribution
-import org.apache.spark.streaming.ui.BatchInfo
 
 /** Base trait for events related to StreamingListener */
 sealed trait StreamingListenerEvent
